refactor(backend): add app factory so it can be used in testing
